pyproject.toml
[project]
name = "mvp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "black>=25.1.0",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
